Add bundlesize to Pipeline
The bundle size of the application is something that is often forgotten
about while in development. To save time of having to go back and strip
away dependancies and reduce code this will be tracked during
development.

Need to find a good threshold to fail under using current for testing.